Create README.md
To use the API, you will need to register for an API key and then make HTTP requests to the API endpoints using Python. You can use the requests library to make HTTP requests and the json library to parse the JSON data returned by the API.

Once you have retrieved the data from the API, you can display it in your Streamlit app using various components such as text, images, tables, and charts. For example, you can display a list of movies based on user input, along with their ratings, release dates, and other details. You can also allow users to search for movies using keywords, and display the results in a user-friendly way.

Overall, the IMDb Movies API is a powerful tool for creating movie-related applications in Streamlit, and it offers a vast amount of data to work with.